refactor: box選択時の色変化関数の場所変更
